LayoutTests/fast/css-grid-layout/grid-item-auto-margins-alignment-vertical-lr-expected.txt

This test checks that aling-self and justify-self properties are not applied when there is auto-margin in the correponding axis. Instead, auto-margin alignment should be applied.

Direction: LTR | Self Aligmment: center | 1 auto-margin

PASS
Direction: LTR | Self Aligmment: start | 4 auto-margin

PASS
Direction: RTL | Self Aligmment: center | 1 auto-margin

PASS
Direction: RTL | Self Aligmment: start | 4 auto-margin

PASS
